feat/LJJ(#114): 폴더 경로 수정
