--- src/components/Blog/Honey/index.ts
@@ -1,1 +1,0 @@
-export { default as Honey } from './Honey';
