--- src/components/Blog/Honey/type.d.ts
@@ -1,11 +1,0 @@
-import { BlogHoneyType } from '@/apis/blog/type';
-
-export type HoneyContentType = BlogHoneyType;
-
-export type BlogCommentType = {
-  content: string;
-};
-
-export type HoneyLeftSideNavProps = Partial<
-  Pick<HoneyContentType, 'id' | 'blogId' | 'userId'>
->;
