--- src/components/Blog/Post/index.ts
@@ -1,1 +1,0 @@
-export { default as Post } from './Post';
