+++ src/components/Blog/Profile/style.ts
@@ -42,3 +42,33 @@ export const CoupleShortIntroduction = styled.div`
     margin: 12px 0px;
   }
 `;
+
+export const SvgContainer = styled.div`
+  padding: 20px;
+  border: 2px dashed ${({ theme }) => theme.border.primary};
+  border-radius: 50%;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+`;
+
+export const UnConnectedProfileBgDiv = styled.div`
+  width: 100%;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  padding: 24px;
+`;
+
+export const UnConnectedInfoWrapper = styled.div`
+  border-radius: 16px;
+  width: 100%;
+  display: flex;
+  flex-direction: column;
+  gap: 16px;
+  justify-content: center;
+  align-items: center;
+  padding: 24px;
+  border: 1px solid ${({ theme }) => theme.border.primary};
+  background-color: ${({ theme }) => theme.bg.tertiary};
+`;
