+++ src/components/Blog/create/index.ts
@@ -1,0 +1,1 @@
+export { default as create } from './Post';
