+++ src/components/Layouts/Profile/index.ts
@@ -1,2 +1,1 @@
-export { default as UnConnectedProfile } from './UnConnected';
 export { default as TogetherImage } from './TogetherImage';
