+++ src/routes/ValidationRoute.tsx
@@ -7,8 +7,8 @@ const SingleBlog = lazy(() => import('@/components/Main/Main'));
 const Blog = lazy(() => import('@/pages/blog/Blog'));
 const PublicBlog = lazy(() => import('@/pages/blog/PublicBlog'));
 const ChangePassword = lazy(() => import('@/components/Auth/ChangePassword'));
-const CreatePost = lazy(() => import('@/components/Blog/Post/Post'));
-const EditPost = lazy(() => import('@/components/Blog/Post/EditBlogPost'));
+const CreatePost = lazy(() => import('@/components/Blog/create/Post'));
+const EditPost = lazy(() => import('@/components/Blog/create/EditBlogPost'));
 const BlogPost = lazy(() => import('@/pages/blog/Post'));
 const Setting = lazy(() => import('@/components/Setting/Setting'));
 const Support = lazy(() => import('@/components/Setting/Support'));
